Add From<Vec<T>> for ParameterizedValue
The impetus behind this change is for creating a derive macro that can translate a struct containing a Vec<T> to an vec of ParameterizedValues.

This also removes `From<Vec<T>> for DatabaseValue` because it had a conflicting type definition. It didn’t appear to be used internally, and it seems easier to work around not having that type.